+++ src/calculator.version.js
@@ -1,1 +1,1 @@
-﻿var CalculatorVersion = "2.0.1.2";
+var CalculatorVersion = "2.1.0.0";